vuln_plugins/subdomain_enumeration.py
def scan(scanner):
    """Scan for Subdomain Enumeration"""
    print("[PLUGIN] Scanning for Subdomain Enumeration...")
    
    try:
        from urllib.parse import urlparse
        parsed_url = urlparse(scanner.target_url)
        domain = parsed_url.netloc
        
        # Remove port if present
        if ':' in domain:
            domain = domain.split(':')[0]
        
        # Common subdomain wordlist
        common_subdomains = [
            'www', 'mail', 'ftp', 'admin', 'blog', 'dev', 'test', 'staging',
            'api', 'app', 'cdn', 'cloud', 'docs', 'help', 'support', 'forum',
            'shop', 'store', 'secure', 'login', 'portal', 'dashboard', 'panel',
            'cpanel', 'webmail', 'ns1', 'ns2', 'dns', 'mx', 'smtp', 'pop',
            'imap', 'vpn', 'remote', 'ssh', 'telnet', 'ftp', 'sftp', 'git',
            'svn', 'jenkins', 'jira', 'confluence', 'wiki', 'redmine', 'mantis',
            'bugzilla', 'trac', 'gitlab', 'github', 'bitbucket', 'sonar',
            'nexus', 'artifactory', 'docker', 'kubernetes', 'rancher', 'prometheus',
            'grafana', 'kibana', 'elasticsearch', 'logstash', 'redis', 'memcached',
            'rabbitmq', 'kafka', 'zookeeper', 'etcd', 'consul', 'vault', 'hashi',
            'terraform', 'packer', 'vagrant', 'ansible', 'chef', 'puppet', 'salt',
            'monitoring', 'alerting', 'backup', 'archive', 'legacy', 'old', 'new',
            'beta', 'alpha', 'rc', 'release', 'production', 'prod', 'development',
            'dev', 'staging', 'stage', 'testing', 'test', 'qa', 'uat', 'preprod',
            'sandbox', 'demo', 'trial', 'free', 'premium', 'enterprise', 'corp',
            'internal', 'external', 'public', 'private', 'secure', 'ssl', 'tls',
            'wildcard', 'default', 'fallback', 'backup', 'replica', 'slave',
            'master', 'primary', 'secondary', 'tertiary', 'failover', 'loadbalancer',
            'lb', 'proxy', 'gateway', 'router', 'firewall', 'ids', 'ips', 'waf',
            'cdn', 'cache', 'static', 'media', 'assets', 'images', 'files',
            'uploads', 'downloads', 'backups', 'logs', 'temp', 'tmp', 'data',
            'database', 'db', 'mysql', 'postgres', 'mongo', 'redis', 'elastic',
            'search', 'index', 'catalog', 'inventory', 'orders', 'payments',
            'billing', 'invoice', 'accounting', 'finance', 'hr', 'human',
            'resources', 'employees', 'users', 'customers', 'clients', 'partners',
            'vendors', 'suppliers', 'contractors', 'consultants', 'advisors',
            'managers', 'directors', 'executives', 'ceo', 'cto', 'cfo', 'cio',
            'cso', 'cpo', 'cmo', 'chro', 'clerk', 'secretary', 'assistant',
            'coordinator', 'specialist', 'analyst', 'engineer', 'developer',
            'programmer', 'designer', 'architect', 'consultant', 'advisor',
            'trainer', 'instructor', 'teacher', 'professor', 'lecturer',
            'researcher', 'scientist', 'technician', 'operator', 'administrator',
            'supervisor', 'manager', 'director', 'executive', 'officer', 'chief',
            'president', 'vice', 'senior', 'junior', 'lead', 'principal', 'staff',
            'associate', 'fellow', 'intern', 'trainee', 'apprentice', 'volunteer',
            'contractor', 'consultant', 'freelancer', 'outsource', 'offshore',
            'onshore', 'remote', 'virtual', 'digital', 'online', 'web', 'mobile',
            'desktop', 'laptop', 'tablet', 'phone', 'smartphone', 'wearable',
            'iot', 'embedded', 'real-time', 'batch', 'stream', 'event', 'message',
            'notification', 'alert', 'warning', 'error', 'debug', 'log', 'trace',
            'profile', 'account', 'user', 'member', 'customer', 'client', 'guest',
            'visitor', 'anonymous', 'public', 'private', 'personal', 'business',
            'corporate', 'enterprise', 'government', 'education', 'healthcare',
            'finance', 'banking', 'insurance', 'retail', 'ecommerce', 'travel',
            'hospitality', 'food', 'restaurant', 'hotel', 'resort', 'spa',
            'fitness', 'gym', 'sports', 'entertainment', 'media', 'news',
            'blog', 'forum', 'social', 'network', 'community', 'group', 'team',
            'organization', 'company', 'corporation', 'business', 'enterprise',
            'startup', 'scaleup', 'unicorn', 'ipo', 'public', 'private', 'family',
            'partnership', 'joint', 'venture', 'alliance', 'merger', 'acquisition',
            'divestiture', 'spin-off', 'subsidiary', 'affiliate', 'branch',
            'office', 'location', 'site', 'facility', 'building', 'campus',
            'headquarters', 'hq', 'main', 'primary', 'secondary', 'regional',
            'local', 'global', 'international', 'worldwide', 'national',
            'federal', 'state', 'provincial', 'municipal', 'city', 'county',
            'district', 'ward', 'precinct', 'zone', 'area', 'region', 'territory',
            'country', 'nation', 'continent', 'hemisphere', 'timezone', 'locale',
            'language', 'culture', 'ethnicity', 'religion', 'politics', 'economy',
            'society', 'community', 'population', 'demographics', 'statistics',
            'analytics', 'metrics', 'kpi', 'performance', 'efficiency', 'quality',
            'safety', 'security', 'privacy', 'compliance', 'governance', 'risk',
            'audit', 'assessment', 'evaluation', 'review', 'inspection', 'test',
            'validation', 'verification', 'certification', 'accreditation',
            'licensing', 'registration', 'enrollment', 'subscription', 'membership',
            'loyalty', 'rewards', 'points', 'credits', 'balance', 'account',
            'wallet', 'payment', 'billing', 'invoice', 'receipt', 'transaction',
            'order', 'purchase', 'sale', 'trade', 'exchange', 'transfer', 'deposit',
            'withdrawal', 'refund', 'chargeback', 'dispute', 'claim', 'appeal',
            'complaint', 'feedback', 'review', 'rating', 'comment', 'message',
            'communication', 'notification', 'alert', 'reminder', 'schedule',
            'calendar', 'appointment', 'meeting', 'conference', 'event', 'webinar',
            'seminar', 'workshop', 'training', 'course', 'class', 'lesson',
            'tutorial', 'guide', 'manual', 'documentation', 'help', 'support',
            'assistance', 'service', 'maintenance', 'repair', 'upgrade', 'update',
            'patch', 'fix', 'bug', 'issue', 'problem', 'error', 'fault', 'failure',
            'crash', 'hang', 'freeze', 'slow', 'performance', 'speed', 'latency',
            'throughput', 'bandwidth', 'capacity', 'storage', 'memory', 'disk',
            'cpu', 'gpu', 'network', 'internet', 'intranet', 'extranet', 'vpn',
            'lan', 'wan', 'wlan', 'bluetooth', 'wifi', 'ethernet', 'fiber',
            'copper', 'wireless', 'cellular', 'satellite', 'radio', 'tv',
            'broadcast', 'stream', 'download', 'upload', 'sync', 'backup',
            'restore', 'recovery', 'disaster', 'business', 'continuity', 'plan',
            'strategy', 'tactic', 'method', 'approach', 'technique', 'tool',
            'technology', 'platform', 'framework', 'library', 'api', 'sdk',
            'plugin', 'extension', 'module', 'component', 'service', 'microservice',
            'function', 'lambda', 'serverless', 'container', 'docker', 'kubernetes',
            'orchestration', 'deployment', 'release', 'version', 'build', 'compile',
            'package', 'bundle', 'archive', 'compression', 'encryption', 'hashing',
            'signing', 'certificate', 'key', 'token', 'password', 'credential',
            'authentication', 'authorization', 'permission', 'role', 'group',
            'user', 'account', 'profile', 'identity', 'session', 'cookie',
            'cache', 'database', 'storage', 'file', 'directory', 'folder',
            'path', 'url', 'uri', 'endpoint', 'route', 'api', 'rest', 'graphql',
            'soap', 'xml', 'json', 'yaml', 'csv', 'excel', 'pdf', 'word',
            'powerpoint', 'image', 'video', 'audio', 'text', 'binary', 'hex',
            'base64', 'url', 'encoding', 'decoding', 'parsing', 'validation',
            'sanitization', 'filtering', 'escaping', 'quoting', 'formatting',
            'serialization', 'deserialization', 'marshalling', 'unmarshalling',
            'conversion', 'transformation', 'migration', 'import', 'export',
            'sync', 'replication', 'backup', 'restore', 'archive', 'compress',
            'decompress', 'encrypt', 'decrypt', 'hash', 'verify', 'sign',
            'verify', 'authenticate', 'authorize', 'validate', 'sanitize',
            'filter', 'escape', 'quote', 'format', 'serialize', 'deserialize',
            'convert', 'transform', 'migrate', 'import', 'export', 'sync',
            'replicate', 'backup', 'restore', 'archive', 'compress', 'decompress',
            'encrypt', 'decrypt', 'hash', 'verify', 'sign', 'verify'
        ]
        
        found_subdomains = []
        
        # Test common subdomains
        for subdomain in common_subdomains:
            test_domain = f"{subdomain}.{domain}"
            test_url = f"{parsed_url.scheme}://{test_domain}"
            
            try:
                response = scanner.session.get(test_url, timeout=scanner.timeout)
                if response.status_code == 200:
                    found_subdomains.append(test_domain)
                    scanner.log_vulnerability(
                        'Subdomain Discovery',
                        'Info',
                        f'Subdomain found: {test_domain}',
                        f'URL: {test_url}, Status: {response.status_code}',
                        'CWE-200'
                    )
            except:
                continue
        
        # DNS enumeration (if dnspython is available)
        try:
            import dns.resolver
            import dns.reversename
            
            # Try to resolve the domain
            try:
                answers = dns.resolver.resolve(domain, 'A')
                scanner.log_vulnerability(
                    'DNS Information',
                    'Info',
                    f'Domain resolves to: {[str(rdata) for rdata in answers]}',
                    f'Domain: {domain}',
                    'CWE-200'
                )
            except:
                pass
            
            # Try common DNS records
            dns_records = ['A', 'AAAA', 'MX', 'NS', 'TXT', 'CNAME', 'SOA']
            for record_type in dns_records:
                try:
                    answers = dns.resolver.resolve(domain, record_type)
                    scanner.log_vulnerability(
                        'DNS Record Exposure',
                        'Low',
                        f'DNS {record_type} record found',
                        f'Domain: {domain}, Record: {record_type}, Values: {[str(rdata) for rdata in answers]}',
                        'CWE-200'
                    )
                except:
                    continue
                    
        except ImportError:
            print("[PLUGIN][Subdomain Enumeration] dnspython not available for DNS enumeration")
        
        # Summary
        if found_subdomains:
            print(f"[PLUGIN][Subdomain Enumeration] Found {len(found_subdomains)} subdomains")
        else:
            print("[PLUGIN][Subdomain Enumeration] No additional subdomains found")
            
    except Exception as e:
        print(f"[PLUGIN][Subdomain Enumeration] Error: {e}") 